+++ pkg/merchantorder/response.go
@@ -28,20 +28,20 @@ type Response struct {
 	RefundedAmount    float64 `json:"refunded_amount"`
 	ShippingCost      float64 `json:"shipping_cost"`
 	TotalAmount       float64 `json:"total_amount"`
-	ID                int64   `json:"id"`
+	ID                int     `json:"id"`
 	Cancelled         bool    `json:"cancelled"`
 }
 
 // PayerResponse represents buyer information.
 type PayerResponse struct {
 	Nickname string `json:"nickname"`
-	ID       int64  `json:"id"`
+	ID       int    `json:"id"`
 }
 
 // CollectorResponse represents collector information.
 type CollectorResponse struct {
 	Nickname string `json:"nickname"`
-	ID       int64  `json:"id"`
+	ID       int    `json:"id"`
 }
 
 // PaymentResponse represents payment information.
@@ -58,7 +58,7 @@ type PaymentResponse struct {
 	TotalPaidAmount   float64 `json:"total_paid_amount"`
 	ShippingCost      float64 `json:"shipping_cost"`
 	AmountRefunded    float64 `json:"amount_refunded"`
-	ID                int64   `json:"id"`
+	ID                int     `json:"id"`
 }
 
 // ItemResponse represents item information.
@@ -87,9 +87,9 @@ type ShipmentResponse struct {
 	Status            string           `json:"status"`
 	ShippingSubstatus string           `json:"shipping_substatus"`
 	ServiceID         string           `json:"service_id"`
-	ID                int64            `json:"id"`
-	SenderID          int64            `json:"sender_id"`
-	ReceiverID        int64            `json:"receiver_id"`
+	ID                int              `json:"id"`
+	SenderID          int              `json:"sender_id"`
+	ReceiverID        int              `json:"receiver_id"`
 	Items             []map[string]any `json:"items"`
 }
 
@@ -110,7 +110,7 @@ type ReceiverAddressResponse struct {
 	Phone        string `json:"phone"`
 	Latitude     string `json:"latitude"`
 	Longitude    string `json:"longitude"`
-	ID           int64  `json:"id"`
+	ID           int    `json:"id"`
 }
 
 // ShippingOptionResponse represents shipping option information.
@@ -122,8 +122,8 @@ type ShippingOptionResponse struct {
 	CurrencyID       string  `json:"currency_id"`
 	Cost             float64 `json:"cost"`
 	ListCost         float64 `json:"list_cost"`
-	ID               int64   `json:"id"`
-	ShippingMethodID int64   `json:"shipping_method_id"`
+	ID               int     `json:"id"`
+	ShippingMethodID int     `json:"shipping_method_id"`
 }
 
 // ReceiverAddressCityResponse represents city information.
@@ -154,6 +154,6 @@ type ShippingEstimatedDeliveryResponse struct {
 
 // ShippingSpeedResponse represents shipping speed information.
 type ShippingSpeedResponse struct {
-	Handling int64 `json:"handling"`
-	Shipping int64 `json:"shipping"`
+	Handling int `json:"handling"`
+	Shipping int `json:"shipping"`
 }
