+++ pkg/payment/client.go
@@ -30,19 +30,19 @@ type Client interface {
 	// Get gets a payment by its ID.
 	// It is a get request to the endpoint: https://api.mercadopago.com/v1/payments/{id}
 	// Reference: https://www.mercadopago.com/developers/en/reference/payments/_payments_id/get/
-	Get(ctx context.Context, id int64) (*Response, error)
+	Get(ctx context.Context, id int) (*Response, error)
 
 	// Cancel cancels a payment by its ID.
 	// It is a put request to the endpoint: https://api.mercadopago.com/v1/payments/{id}
-	Cancel(ctx context.Context, id int64) (*Response, error)
+	Cancel(ctx context.Context, id int) (*Response, error)
 
 	// Capture captures a payment by its ID.
 	// It is a put request to the endpoint: https://api.mercadopago.com/v1/payments/{id}
-	Capture(ctx context.Context, id int64) (*Response, error)
+	Capture(ctx context.Context, id int) (*Response, error)
 
 	// CaptureAmount captures amount of a payment by its ID.
 	// It is a put request to the endpoint: https://api.mercadopago.com/v1/payments/{id}
-	CaptureAmount(ctx context.Context, id int64, amount float64) (*Response, error)
+	CaptureAmount(ctx context.Context, id int, amount float64) (*Response, error)
 }
 
 // client is the implementation of Client.
@@ -87,7 +87,7 @@ func (c *client) Search(ctx context.Context, request SearchRequest) (*SearchResp
 	return result, nil
 }
 
-func (c *client) Get(ctx context.Context, id int64) (*Response, error) {
+func (c *client) Get(ctx context.Context, id int) (*Response, error) {
 	pathParams := map[string]string{
 		"id": strconv.Itoa(int(id)),
 	}
@@ -105,7 +105,7 @@ func (c *client) Get(ctx context.Context, id int64) (*Response, error) {
 	return result, nil
 }
 
-func (c *client) Cancel(ctx context.Context, id int64) (*Response, error) {
+func (c *client) Cancel(ctx context.Context, id int) (*Response, error) {
 	request := &CancelRequest{Status: "cancelled"}
 
 	pathParams := map[string]string{
@@ -126,7 +126,7 @@ func (c *client) Cancel(ctx context.Context, id int64) (*Response, error) {
 	return result, nil
 }
 
-func (c *client) Capture(ctx context.Context, id int64) (*Response, error) {
+func (c *client) Capture(ctx context.Context, id int) (*Response, error) {
 	request := &CaptureRequest{Capture: true}
 
 	pathParams := map[string]string{
@@ -147,7 +147,7 @@ func (c *client) Capture(ctx context.Context, id int64) (*Response, error) {
 	return result, nil
 }
 
-func (c *client) CaptureAmount(ctx context.Context, id int64, amount float64) (*Response, error) {
+func (c *client) CaptureAmount(ctx context.Context, id int, amount float64) (*Response, error) {
 	request := &CaptureRequest{TransactionAmount: amount, Capture: true}
 
 	pathParams := map[string]string{
